Refactor `svanna-benchmark` content
