+++ svanna-benchmark/src/main/java/org/jax/svanna/benchmark/Main.java
@@ -1,5 +1,6 @@
 package org.jax.svanna.benchmark;
 
+import org.jax.svanna.benchmark.cmd.benchmark_case.BenchmarkCaseCommand;
 import picocli.CommandLine;
 
 import java.util.Locale;
