+++ svanna-benchmark/src/main/java/org/jax/svanna/benchmark/cmd/benchmark_case/BenchmarkCaseCommand.java
@@ -1,8 +1,12 @@
-package org.jax.svanna.benchmark;
+package org.jax.svanna.benchmark.cmd.benchmark_case;
 
 import org.apache.commons.compress.compressors.gzip.GzipCompressorOutputStream;
 import org.apache.commons.csv.CSVFormat;
 import org.apache.commons.csv.CSVPrinter;
+import org.jax.svanna.benchmark.cmd.BaseBenchmarkCommand;
+import org.jax.svanna.benchmark.Main;
+import org.jax.svanna.benchmark.io.CaseReport;
+import org.jax.svanna.benchmark.io.CaseReportImporter;
 import org.jax.svanna.benchmark.util.ProgressReporter;
 import org.jax.svanna.benchmark.util.TaskUtils;
 import org.jax.svanna.core.LogUtils;
