+++ svanna-benchmark/src/test/resources/org/jax/svanna/benchmark/io/Iqbal-2013-ANK3-V:1.json
