+++ svanna-benchmark/src/test/resources/org/jax/svanna/benchmark/io/Witzel-2017-SMARCD2-BII.1.json
